Trait oak_grpc::oak::containers::launcher_server::Launcher

pub trait Launcher:
    Send
    + Sync
    + 'static {
    type GetOakSystemImageStream: Stream<Item = Result<GetImageResponse, Status>> + Send + 'static;
    type GetContainerBundleStream: Stream<Item = Result<GetImageResponse, Status>> + Send + 'static;

    // Required methods
    fn get_oak_system_image<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetOakSystemImageStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_container_bundle<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetContainerBundleStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_application_config<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetApplicationConfigResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_endorsements<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Endorsements>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_attestation_evidence<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SendAttestationEvidenceRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn notify_app_ready<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with LauncherServer.

Required Associated Types

type GetOakSystemImageStream: Stream<Item = Result<GetImageResponse, Status>> + Send + 'static

Server streaming response type for the GetOakSystemImage method.

type GetContainerBundleStream: Stream<Item = Result<GetImageResponse, Status>> + Send + 'static

Server streaming response type for the GetContainerBundle method.

Required Methods

fn get_oak_system_image<'life0, 'async_trait>( &'life0 self, request: Request<()>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetOakSystemImageStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provides stage1 with the Oak system image (which contains the Linux distribution and the orchestrator binary).

fn get_container_bundle<'life0, 'async_trait>( &'life0 self, request: Request<()>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetContainerBundleStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provides orchestrator with the trusted container image.

fn get_application_config<'life0, 'async_trait>( &'life0 self, request: Request<()>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetApplicationConfigResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

This method is used by the orchestrator to load and measure the trusted application config. The orchestrator will later, separately expose this config to the application.

fn get_endorsements<'life0, 'async_trait>( &'life0 self, request: Request<()>, ) -> Pin<Box<dyn Future<Output = Result<Response<Endorsements>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provides the orchestrator with the endorsements of the trusted application and the container.

fn send_attestation_evidence<'life0, 'async_trait>( &'life0 self, request: Request<SendAttestationEvidenceRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sends Attestation Evidence containing the Attestation Report with corresponding measurements and public keys to the Launcher. This API is called exactly once after the Attestation Evidence is generated. Calling this API a second time will result in an error.

fn notify_app_ready<'life0, 'async_trait>( &'life0 self, request: Request<()>, ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Notifies the launcher that the trusted app is ready to serve requests and listening on the pre-arranged port (8080).