Trait oak_grpc::oak::session::v1::streaming_session_server::StreamingSession

pub trait StreamingSession:
    Send
    + Sync
    + 'static {
    type StreamStream: Stream<Item = Result<ResponseWrapper, Status>> + Send + 'static;

    // Required method
    fn stream<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<RequestWrapper>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with StreamingSessionServer.

Required Associated Types

type StreamStream: Stream<Item = Result<ResponseWrapper, Status>> + Send + 'static

Server streaming response type for the Stream method.

Required Methods

fn stream<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<RequestWrapper>>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Used to send a sequence of messages ensuring that they are all handled by the same server instance, by virtue of being multiplexed over a single gRPC stream.

The RequestWrapper and ResponseWrapper messages are thin wrappers around the underlying messages exchanged by client and server, giving it a minimal amount of structure and type safety.

The expected message sequence starts with the client sending a GetEndorsedEvidenceRequest message in order to fetch the evidence of the enclave. This method may be handled by the untrusted launcher or by the enclave, depending on the server implementation.

Then the client encrypts the payload with the public key contained in the evidence via a hybrid encryption protocol, and sends the encrypted payload as part of a InvokeRequest message.